Updated/Fixed Delete exception

+++ src/main/java/com/godknows/crudchallenge/services/ClientService.java
@@ -3,15 +3,17 @@ package com.godknows.crudchallenge.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.godknows.crudchallenge.DTOs.ClientDTO;
 import com.godknows.crudchallenge.entities.Client;
 import com.godknows.crudchallenge.repositories.ClientRepository;
+import com.godknows.crudchallenge.services.exceptions.DatabaseException;
 import com.godknows.crudchallenge.services.exceptions.ResourceNotFoundException;
 
 import jakarta.persistence.EntityNotFoundException;
@@ -57,13 +59,16 @@ public class ClientService {
 
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
+		 if(!clientRepo.existsById(id)) {
+			 throw new ResourceNotFoundException("Cliente não encontrado");
+		 }
 		try {
 			clientRepo.deleteById(id);
 		}
-		catch (EmptyResultDataAccessException e) {
-			throw new ResourceNotFoundException("Recuso não encontrado");
+		catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Recuso não encontrado");
 		}
 	}
